fix compile in iar

+++ src/mqtt/mqtt_example.c
@@ -8,11 +8,11 @@
 void *HAL_Malloc(uint32_t size);
 void HAL_Free(void *ptr);
 void HAL_Printf(const char *fmt, ...);
-int8_t HAL_GetProductKey(char product_key[IOTX_PRODUCT_KEY_LEN]);
-int8_t HAL_GetDeviceName(char device_name[IOTX_DEVICE_NAME_LEN]);
+int HAL_GetProductKey(char product_key[IOTX_PRODUCT_KEY_LEN]);
+int HAL_GetDeviceName(char device_name[IOTX_DEVICE_NAME_LEN]);
 int HAL_GetDeviceSecret(char device_secret[IOTX_DEVICE_SECRET_LEN]);
 uint64_t HAL_UptimeMs(void);
-int8_t HAL_Snprintf(char *str, const int len, const char *fmt, ...);
+int HAL_Snprintf(char *str, const int len, const char *fmt, ...);
 
 #define EXAMPLE_TRACE(fmt, ...)  \
     do { \
