+++ src/mqtt/mqtt_example_at.c
@@ -30,10 +30,10 @@ extern char _device_secret[IOTX_DEVICE_SECRET_LEN + 1];
 void *HAL_Malloc(uint32_t size);
 void HAL_Free(void *ptr);
 void HAL_Printf(const char *fmt, ...);
-int8_t HAL_GetProductKey(char product_key[IOTX_PRODUCT_KEY_LEN]);
-int8_t HAL_GetDeviceName(char device_name[IOTX_DEVICE_NAME_LEN]);
+int HAL_GetProductKey(char product_key[IOTX_PRODUCT_KEY_LEN]);
+int HAL_GetDeviceName(char device_name[IOTX_DEVICE_NAME_LEN]);
 uint64_t HAL_UptimeMs(void);
-int8_t HAL_Snprintf(char *str, const int len, const char *fmt, ...);
+int HAL_Snprintf(char *str, const int len, const char *fmt, ...);
 void HAL_SleepMs(uint32_t ms);
 
 #define EXAMPLE_TRACE(fmt, ...)  \
